+++ packages/deepsea-tools/src/utils/writeExcel.ts
@@ -1,28 +1,36 @@
-import { utils, writeFile } from "xlsx"
+import { JSON2SheetOpts, utils, writeFile } from "xlsx"
 
-export type WriteExcelItem = string | number | Date | boolean | undefined | null
+export { JSON2SheetOpts } from "xlsx"
 
-export type WriteExcelData = Record<string, WriteExcelItem>[]
+export type WriteableWorkSheetCellData = string | number | Date | boolean | undefined | null
+
+export type WriteableWorkSheetData = Record<string, WriteableWorkSheetCellData>[]
+
+export type WriteableWorkBookData = Record<string, WriteableWorkSheetData>
 
 export interface WriteExcelParams {
     /**
      * 导出的数据
      */
-    data: WriteExcelData | Record<string, WriteExcelData>
+    data: WriteableWorkSheetData | WriteableWorkBookData
     /**
      * 导出的文件名
      */
     path: string
+    /**
+     * 导出的选项
+     */
+    jsonToSheetOptions?: JSON2SheetOpts
 }
 
 const defaultSheetName = "__DEFAULT_SHEET_NAME__"
 
 /** 手动导出 excel */
-export function writeExcel({ data, path }: WriteExcelParams) {
+export function writeExcel({ data, path, jsonToSheetOptions }: WriteExcelParams) {
     data = Array.isArray(data) ? { [defaultSheetName]: data } : data
     const workBook = utils.book_new()
     Object.entries(data).forEach(([key, value]) => {
-        const workSheet = utils.json_to_sheet(value)
+        const workSheet = utils.json_to_sheet(value, jsonToSheetOptions)
         utils.book_append_sheet(workBook, workSheet, key === defaultSheetName ? undefined : key)
     })
     writeFile(workBook, path.replace(/^(.+?)\.xlsx?$/i, "$1.xlsx"))