+++ packages/soda-antd/CHANGELOG.md
@@ -1,5 +1,13 @@
 # soda-antd
 
+## 0.4.12
+
+### Patch Changes
+
+- Updated dependencies
+    - deepsea-tools@5.24.1
+    - soda-hooks@6.6.4
+
 ## 0.4.11
 
 ### Patch Changes